update copyright block in all files and rename edge_detection -> edges
