--- curve_apps/edge_detection/__init__.py
@@ -1,5 +1,0 @@
-#  Copyright (c) 2024 Mira Geoscience Ltd.
-#
-#  This file is part of curve-apps package.
-#
-#  All rights reserved.
