+++ curve_apps/edges/__init__.py
@@ -1,7 +1,10 @@
 #  '''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
 #  Copyright (c) 2024 Mira Geoscience Ltd.                                       '
 #                                                                                '
+#  This file is part of edges package.                                        '
+#                                                                                '
 #  All rights reserved.                                                          '
+#                                                                                '
 #                                                                                '
 #  This file is part of curve-apps.                                              '
 #                                                                                '
